backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
scikit-image
python-pptx
tqdm
aiofiles